update
Lista de Rolos e exposições a funcionar
Add rolo a funcionar

Falta:
Verificações de base de dados vazia (esta a estoirar)
Verificações de Formulario

+++ app/src/main/java/pt/isec/gps/grupo14/analog/AnaLog/Rolo.java
@@ -21,7 +21,7 @@ public class Rolo {
     private int maxExposicoes;
     private int nExposicoes;
 
-    public Rolo(String titulo, float iso, int formato, int maxExposicoes, String descricao,
+    public Rolo(String titulo, int iso, int formato, int maxExposicoes, String descricao,
                 int idCamera, Context context) {
         this.titulo = titulo;
         this.iso = iso;
